+++ packages/curd-table/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @naive-potato-ui/curd-table
 
+## 1.0.9
+
+### Patch Changes
+
+- fix query
+
 ## 1.0.8
 
 ### Patch Changes